`application.properties` renommé en `application.properties.sav` pour pourvoir jouer les tests d'intération J'ai recopié config/application.properties dans src/main/resources/config pour un demarrage de l'application en dev de façon similaire
SpringApplication will load properties from application.properties files in the following locations and add them to the Spring Environment:

A /config subdirectory of the current directory.
The current directory
A classpath /config package
The classpath root
The list is ordered by precedence (properties defined in locations higher in the list override those defined in lower locations).